General refactoring
- Changing package conf to basic_data.
