+++ impc_prod_tracker/core/src/main/java/org/gentar/basic_data/CatalogServiceImpl.java
@@ -13,7 +13,7 @@
  * language governing permissions and limitations under the
  * License.
  *******************************************************************************/
-package org.gentar.conf;
+package org.gentar.basic_data;
 
 
 import org.gentar.biology.plan.attempt.crispr.nuclease.nuclease_class.NucleaseClassRepository;
