+++ impc_prod_tracker/core/src/main/java/org/gentar/basic_data/RestConfiguration.java
@@ -1,4 +1,4 @@
-package org.gentar.conf;
+package org.gentar.basic_data;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
